small rfks, finish cmd
