+++ src/commands/index.ts
@@ -1,5 +1,6 @@
 import pull from "./pull";
 import fetch from "./fetch";
+import extract from "./extract";
 import stop from "./stop";
 import list from "./list";
 import serve from "./serve";
@@ -8,6 +9,7 @@ import clear from "./clear";
 export default {
   pull,
   fetch,
+  extract,
   stop,
   list,
   serve,
